auto-twit/requirements.txt
python-twitter
firebase-admin
apscheduler